# ritter_db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.26.0

package ritter_db

import (
	"database/sql"
)

type Post struct {
	ID        int64
	Content   string
	UserID    int64
	Createdat sql.NullTime
}

type User struct {
	ID        int64
	Nickname  string
	Passwd    string
	Createdat sql.NullTime
}

type UserDatum struct {
	ID     int64
	UserID sql.NullInt64
}